// src/app/services/models/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserDto {
  address?: string;
  adminCode?: string;
  amount?: number;
  createdAt?: string;
  email?: string;
  id?: number;
  latitude?: number;
  longitude?: number;
  name?: string;
  password?: string;
  phoneNum?: string;
  referenceCode?: string;
  userTypes?: 'HORNY_GUY' | 'DATE_GIRL' | 'ADMIN';
}